Rename of modules to avoid broken code
